backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
nltk==3.8.1
textstat==0.7.3
vaderSentiment==3.3.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.4
python-decouple==3.8
# Add these for production deployment:
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9